Add bonding and vlan tagging doc examples
The net.toml version is bumped to 3 to accomidate the new functionality
of bonding and vlans. Version 3 does require different naming than
version 2 so version net.toml files are not drop in ready for version 3.
This also specifies which options are support for bonding and vlans.

Bonding is a complex topic and right now, we are making a deliberate
choice to support only a subset of bonding before full support is
enabled. This mostly comes in the form of only mode 1 (active-primary).

VLAN tagging is supported by creating a virtual network device on top of
another device. This allows mixing and matching of vlan tags with
whatever configuration is required. A user can chose to only configure
VLAN Tagged devices for addressing if that is required, or add in vlan
tagging on a device for special handling.