[packages] Add brcm-firmware package for Raspberry Pi
This currently takes the Broadcom firmware provided via firmware-nonfree
and places it in the firmware location. This is needed to get WiFi
devices working. This still needs other things like regulatory.db and a
wireless network manager like iw or wpa_supplicant. The tarfile should
also be generated and put into the cache. Right now its not provided
until we have a good place for it plus a good way to source these files.