buildsys: Move CA Certificate Bundle out of packages
CA Bundles are a root of trust and can be swapped out similarly to
other roots of trust. This commit moves the fetching/building of the CA
Bundle to a more general location. This change lays the foundation for
future work which will build upon this to allow roots of trust to be
replaced while preserving the rest of the image's functionality. This
will enable new ways to test and vend Bottlerocket images without
building completely from source.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>